+++ measure_test.go
@@ -7,7 +7,13 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func _measure3(node *Node, width float32, widthMode MeasureMode, height float32, heightMode MeasureMode) Size {
+func _measure3(
+	node *Node,
+	width float64,
+	widthMode MeasureMode,
+	height float64,
+	heightMode MeasureMode,
+) Size {
 	measureCount, ok := node.Context.(*int)
 	if ok {
 		(*measureCount)++
@@ -16,7 +22,13 @@ func _measure3(node *Node, width float32, widthMode MeasureMode, height float32,
 	return Size{Width: 10, Height: 10}
 }
 
-func _simulate_wrapping_text(node *Node, width float32, widthMode MeasureMode, height float32, heightMode MeasureMode) Size {
+func _simulate_wrapping_text(
+	node *Node,
+	width float64,
+	widthMode MeasureMode,
+	height float64,
+	heightMode MeasureMode,
+) Size {
 	if widthMode == MeasureModeUndefined || width >= 68 {
 		return Size{Width: 68, Height: 16}
 	}
@@ -24,7 +36,13 @@ func _simulate_wrapping_text(node *Node, width float32, widthMode MeasureMode, h
 	return Size{Width: 50, Height: 32}
 }
 
-func _measure_assert_negative(node *Node, width float32, widthMode MeasureMode, height float32, heightMode MeasureMode) Size {
+func _measure_assert_negative(
+	node *Node,
+	width float64,
+	widthMode MeasureMode,
+	height float64,
+	heightMode MeasureMode,
+) Size {
 	if width < 0 {
 		panic(fmt.Sprintf("width is %.2f and should be >= 0", width))
 	}
